@@ -1,26 +1,25 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import styled from 'styled-components';
-import SearchState from '../models/ui/SearchState';
-import { localize, pluralize } from '../utils/formatters';
-import { connect } from '../utils/helpers';
-import { orgDetailsPath } from '../utils/urls';
-import NewOrgForm from './orgs/NewOrgForm';
-import NewProjectForm from './projects/NewProjectForm';
-import ProjectContent from './projects/ProjectContent';
-import AppBar from './ui/AppBar';
-import { Button, RaisedButton } from './ui/Buttons';
-import { FlexBetween, FlexCenter } from './ui/Flexboxes';
-import { PageHeader } from './ui/Headers';
-import { ArrowDownIcon } from './ui/Icons';
-import { Link } from './ui/Links';
-import Loading from './ui/Loading';
-import Logo from './ui/Logo';
-import Menu from './ui/Menu';
-import Search from './ui/Search';
-import { Section } from './ui/Sections';
-import { Heading, PageTitle, SectionTitle, Text } from './ui/Texts';
-import { Width640 } from './ui/Widths';
+import React, { useContext, useState } from "react";
+import styled from "styled-components";
+import { listOrgs, listProjects } from "../api";
+import { flatMap, localize, pluralize } from "../utils";
+import Context from "./Context";
+import useForm from "./hooks/useForm";
+import useLoading from "./hooks/useLoading";
+import NewOrgForm from "./orgs/NewOrgForm";
+import NewProjectForm from "./projects/NewProjectForm";
+import ProjectContent from "./projects/ProjectContent";
+import AppBar from "./ui/AppBar";
+import { Button, RaisedButton } from "./ui/Buttons";
+import { FlexBetween } from "./ui/Flexboxes";
+import { PageHeader } from "./ui/Headers";
+import { ArrowDownIcon } from "./ui/Icons";
+import { Link } from "./ui/Links";
+import Logo from "./ui/Logo";
+import Menu from "./ui/Menu";
+import Search from "./ui/Search";
+import { Section } from "./ui/Sections";
+import { Heading, PageTitle, SectionTitle, Text } from "./ui/Texts";
+import { Width640 } from "./ui/Widths";
 
 const WelcomeHeader = styled(PageHeader)`
   background-color: var(--color-black);
@@ -43,34 +42,49 @@ const NewButtonIcon = styled(ArrowDownIcon)`
   width: var(--size-16);
 `;
 
-const LoadingWrapper = styled(FlexCenter)`
-  margin-bottom: var(--size-32);
-`;
+export default function WelcomePage() {
+  const { openDrawer } = useContext(Context);
 
-class WelcomePage extends React.Component {
-  search = SearchState.create();
+  const [orgs, setOrgs] = useState([]);
+  const [projects, setProjects] = useState([]);
 
-  render() {
-    const { isLoading, openDrawer, orgs } = this.props;
-    return (
-      <main>
-        <WelcomeHeader>
-          <Width640>
-            <AppBar>
-              <Logo />
-            </AppBar>
-            <Title>Welcome!</Title>
-            <Heading>
-              Use the BitBrew Console to manage your data pipeline.
-            </Heading>
-          </Width640>
-        </WelcomeHeader>
+  async function loadOrgs() {
+    const { items } = await listOrgs();
+    const projectData = await Promise.all(
+      items.map(item => listProjects(item.id)),
+    );
+    setOrgs(items);
+    setProjects(
+      flatMap(projectData, ({ items: projectItems }) => projectItems),
+    );
+  }
+
+  const isLoading = useLoading(loadOrgs);
+
+  const [values, setValue] = useForm({
+    searchTerm: "",
+  });
+
+  return (
+    <main>
+      <WelcomeHeader>
+        <Width640>
+          <AppBar>
+            <Logo />
+          </AppBar>
+          <Title>Welcome!</Title>
+          <Heading>
+            Use the BitBrew Console to manage your data pipeline.
+          </Heading>
+        </Width640>
+      </WelcomeHeader>
+      {!isLoading && (
         <Width640>
           <Actions>
             <Search
               description="The list of organizations below will change to show only those with project names matching the search term."
-              value={this.search.term}
-              onChange={this.search.setTerm}
+              value={values.searchTerm}
+              onChange={setValue}
               placeholder="Search by project name"
             />
             <Menu
@@ -82,70 +96,57 @@ class WelcomePage extends React.Component {
             >
               <Button
                 onClick={() => {
-                  openDrawer(<NewOrgForm />);
+                  openDrawer(<NewOrgForm onCreate={loadOrgs} />);
                 }}
-                disabled={isLoading}
               >
                 Organization
               </Button>
               <Button
                 onClick={() => {
-                  openDrawer(<NewProjectForm />);
+                  openDrawer(<NewProjectForm org={orgs} onCreate={loadOrgs} />);
                 }}
               >
                 Project
               </Button>
             </Menu>
           </Actions>
-          {isLoading && (
-            <LoadingWrapper>
-              <Loading />
-            </LoadingWrapper>
-          )}
           {orgs.map(org => {
-            const projects = org.getProjectsWithName(this.search.term);
-            const isVisible = this.search.isEmpty || projects.length > 0;
+            const searchResults = projects.filter(
+              project =>
+                project.orgId === org.id &&
+                project.name
+                  .toUpperCase()
+                  .includes(values.searchTerm.toUpperCase()),
+            );
             return (
-              isVisible && (
+              (values.searchTerm === "" || searchResults.length > 0) && (
                 <Section key={org.id}>
                   <SectionTitle>
-                    <Link to={orgDetailsPath(org.id)}>{org.name}</Link>
+                    <Link to={`/orgs/${org.id}`}>{org.name}</Link>
                   </SectionTitle>
                   <FlexBetween>
                     <Text gray>Created on {localize(org.createdAt)}</Text>
-                    <Text gray>{pluralize('project', projects.length)}</Text>
+                    <Text gray>
+                      {pluralize("project", searchResults.length)}
+                    </Text>
                   </FlexBetween>
-                  {projects.map(project => (
-                    <ProjectContent key={project.id} project={project} />
+                  {searchResults.map(project => (
+                    <ProjectContent
+                      key={project.id}
+                      project={{
+                        ...project,
+                        orgName: org.name,
+                      }}
+                      onUpdate={loadOrgs}
+                      onDelete={loadOrgs}
+                    />
                   ))}
                 </Section>
               )
             );
           })}
         </Width640>
-      </main>
-    );
-  }
+      )}
+    </main>
+  );
 }
-
-WelcomePage.propTypes = {
-  isLoading: PropTypes.bool.isRequired,
-  openDrawer: PropTypes.func.isRequired,
-  orgs: PropTypes.arrayOf(
-    PropTypes.shape({
-      createdAt: PropTypes.string.isRequired,
-      getProjectsWithName: PropTypes.func.isRequired,
-      id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
-    }),
-  ).isRequired,
-};
-
-export default connect(
-  WelcomePage,
-  ({ orgStore, uiStore }) => ({
-    isLoading: orgStore.isLoading,
-    openDrawer: uiStore.openDrawer,
-    orgs: orgStore.orgs,
-  }),
-);
