@@ -1,17 +1,17 @@
 module.exports = function buildBabelPreset() {
-  const isDevelopmentEnv = process.env.NODE_ENV === 'development';
-  const isProductionEnv = process.env.NODE_ENV === 'production';
+  const isDevelopmentEnv = process.env.NODE_ENV === "development";
+  const isProductionEnv = process.env.NODE_ENV === "production";
   return {
     presets: [
       [
-        '@babel/preset-env',
+        "@babel/preset-env",
         {
           modules: false,
-          useBuiltIns: 'usage',
+          useBuiltIns: "usage",
         },
       ],
       [
-        '@babel/preset-react',
+        "@babel/preset-react",
         {
           development: isDevelopmentEnv,
           useBuiltIns: true,
@@ -19,17 +19,15 @@ module.exports = function buildBabelPreset() {
       ],
     ],
     plugins: [
-      '@babel/plugin-proposal-class-properties',
-      '@babel/plugin-proposal-optional-catch-binding',
       [
-        'babel-plugin-styled-components',
+        "babel-plugin-styled-components",
         {
           displayName: isDevelopmentEnv,
           pure: true,
         },
       ],
       isProductionEnv && [
-        'babel-plugin-transform-react-remove-prop-types',
+        "babel-plugin-transform-react-remove-prop-types",
         {
           removeImport: true,
         },
