@@ -1,20 +1,22 @@
-import { navigate } from '@reach/router';
-import PropTypes from 'prop-types';
-import React from 'react';
-import { connect } from '../../utils/helpers';
-import { projectDevicesPath } from '../../utils/urls';
-import DeleteDialog from '../ui/DeleteDialog';
-import Panel from '../ui/Panel';
-import { Text } from '../ui/Texts';
+import PropTypes from "prop-types";
+import React from "react";
+import { deleteDevice, listDevices } from "../../api";
+import { poll } from "../../utils";
+import DeleteDialog from "../ui/DeleteDialog";
+import Panel from "../ui/Panel";
+import { Text } from "../ui/Texts";
 
-function DeleteDeviceDialog({ deleteDevice, device }) {
+export default function DeleteDeviceDialog({ device, onDelete }) {
   return (
     <DeleteDialog
       title="Delete Device"
       onDelete={async () => {
-        const projectId = device.project.id;
-        await deleteDevice(device);
-        navigate(projectDevicesPath(projectId));
+        await deleteDevice(device.id);
+        await poll(async () => {
+          const { items } = await listDevices(device.projectId);
+          return items.every(item => item.id !== device.id);
+        });
+        await onDelete();
       }}
     >
       <Text>
@@ -23,9 +25,9 @@ function DeleteDeviceDialog({ deleteDevice, device }) {
       </Text>
       <Panel
         items={[
-          ['Device', device.codename],
-          ['Project', device.project.name],
-          ['Organization', device.project.org.name],
+          ["Device", device.codename],
+          ["Project", device.projectName],
+          ["Organization", device.orgName],
         ]}
       />
       <Text>Are you sure you want to continue?</Text>
@@ -34,22 +36,12 @@ function DeleteDeviceDialog({ deleteDevice, device }) {
 }
 
 DeleteDeviceDialog.propTypes = {
-  deleteDevice: PropTypes.func.isRequired,
+  onDelete: PropTypes.func.isRequired,
   device: PropTypes.shape({
     codename: PropTypes.string.isRequired,
-    project: PropTypes.shape({
-      name: PropTypes.string.isRequired,
-      id: PropTypes.string.isRequired,
-      org: PropTypes.shape({
-        name: PropTypes.string.isRequired,
-      }).isRequired,
-    }).isRequired,
+    id: PropTypes.string.isRequired,
+    orgName: PropTypes.string.isRequired,
+    projectId: PropTypes.string.isRequired,
+    projectName: PropTypes.string.isRequired,
   }).isRequired,
 };
-
-export default connect(
-  DeleteDeviceDialog,
-  ({ deviceStore }) => ({
-    deleteDevice: deviceStore.deleteDevice,
-  }),
-);
