@@ -1,34 +1,38 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import FormState from '../../models/ui/FormState';
-import { connect } from '../../utils/helpers';
-import FormDrawer from '../ui/FormDrawer';
-import DeviceFormFields from './DeviceFormFields';
+import PropTypes from "prop-types";
+import React from "react";
+import { listDevices, updateDevice } from "../../api";
+import { poll } from "../../utils";
+import useForm from "../hooks/useForm";
+import FormDrawer from "../ui/FormDrawer";
+import DeviceFormFields from "./DeviceFormFields";
 
-class EditDeviceForm extends React.Component {
-  /* eslint-disable react/destructuring-assignment */
-  form = FormState.props({
-    codename: this.props.device.codename,
-    type: this.props.device.type,
-    serialNumber: this.props.device.serialNumber,
-    imei: this.props.device.imei,
-    enabled: this.props.device.enabled,
-    projectId: this.props.device.project.id,
-  }).create();
+export default function EditDeviceForm({ device, onUpdate }) {
+  const [values, setValue] = useForm({
+    codename: device.codename,
+    type: device.type,
+    serialNumber: device.serialNumber,
+    imei: device.imei,
+    enabled: device.enabled,
+    projectId: device.projectId,
+  });
 
-  /* eslint-enable react/destructuring-assignment */
-  render() {
-    const { device, updateDevice } = this.props;
-    return (
-      <FormDrawer
-        title="Edit Device"
-        buttonText="Save"
-        onSubmit={() => updateDevice(device, this.form.serialized)}
-      >
-        <DeviceFormFields form={this.form} />
-      </FormDrawer>
-    );
-  }
+  return (
+    <FormDrawer
+      title="Edit Device"
+      buttonText="Save"
+      onSubmit={async () => {
+        const data = await updateDevice(device.id, values);
+        await poll(async () => {
+          const { items } = await listDevices(data.projectId);
+          const found = items.find(item => item.id === data.id);
+          return !found || found.updatedAt === data.updatedAt;
+        });
+        await onUpdate();
+      }}
+    >
+      <DeviceFormFields values={values} setValue={setValue} />
+    </FormDrawer>
+  );
 }
 
 EditDeviceForm.propTypes = {
@@ -36,18 +40,9 @@ EditDeviceForm.propTypes = {
     codename: PropTypes.string.isRequired,
     enabled: PropTypes.bool.isRequired,
     imei: PropTypes.string.isRequired,
-    project: PropTypes.shape({
-      id: PropTypes.string.isRequired,
-    }).isRequired,
+    projectId: PropTypes.string.isRequired,
     serialNumber: PropTypes.string.isRequired,
     type: PropTypes.string.isRequired,
   }).isRequired,
-  updateDevice: PropTypes.func.isRequired,
+  onUpdate: PropTypes.func.isRequired,
 };
-
-export default connect(
-  EditDeviceForm,
-  ({ deviceStore }) => ({
-    updateDevice: deviceStore.updateDevice,
-  }),
-);
