@@ -1,46 +1,40 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import FormState from '../../models/ui/FormState';
-import { connect } from '../../utils/helpers';
-import FormDrawer from '../ui/FormDrawer';
-import DeviceFormFields from './DeviceFormFields';
+import PropTypes from "prop-types";
+import React from "react";
+import { createDevice, listDevices } from "../../api";
+import { poll } from "../../utils";
+import useForm from "../hooks/useForm";
+import FormDrawer from "../ui/FormDrawer";
+import DeviceFormFields from "./DeviceFormFields";
 
-class NewDeviceForm extends React.Component {
-  /* eslint-disable react/destructuring-assignment */
-  form = FormState.props({
-    codename: '',
-    type: 'Datalogger',
-    serialNumber: '',
-    imei: '',
+export default function NewDeviceForm({ onCreate, project }) {
+  const [values, setValue] = useForm({
+    codename: "",
+    type: "Datalogger",
+    serialNumber: "",
+    imei: "",
     enabled: true,
-    projectId: this.props.project.id,
-  }).create();
+    projectId: project,
+  });
 
-  /* eslint-enable react/destructuring-assignment */
-  render() {
-    const { createDevice } = this.props;
-    return (
-      <FormDrawer
-        title="New Device"
-        buttonText="Create"
-        onSubmit={() => createDevice(this.form.serialized)}
-      >
-        <DeviceFormFields form={this.form} />
-      </FormDrawer>
-    );
-  }
+  return (
+    <FormDrawer
+      title="New Device"
+      buttonText="Create"
+      onSubmit={async () => {
+        const data = await createDevice(values);
+        await poll(async () => {
+          const { items } = await listDevices(data.projectId);
+          return items.some(item => item.id === data.id);
+        });
+        await onCreate();
+      }}
+    >
+      <DeviceFormFields values={values} setValue={setValue} />
+    </FormDrawer>
+  );
 }
 
 NewDeviceForm.propTypes = {
-  createDevice: PropTypes.func.isRequired,
-  project: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-  }).isRequired,
+  onCreate: PropTypes.func.isRequired,
+  project: PropTypes.string.isRequired,
 };
-
-export default connect(
-  NewDeviceForm,
-  ({ deviceStore }) => ({
-    createDevice: deviceStore.createDevice,
-  }),
-);
