@@ -1,22 +1,19 @@
-import { observer } from 'mobx-react';
-import PropTypes from 'prop-types';
-import React from 'react';
-import { Input, Label } from '../ui/Forms';
+import PropTypes from "prop-types";
+import React from "react";
+import { Input, Label } from "../ui/Forms";
 
-function OrgFormFields({ form }) {
+export default function OrgFormFields({ setValue, values }) {
   return (
     <Label htmlFor="name">
       Name
-      <Input id="name" value={form.name} onChange={form.setValue} />
+      <Input id="name" value={values.name} onChange={setValue} />
     </Label>
   );
 }
 
 OrgFormFields.propTypes = {
-  form: PropTypes.shape({
+  setValue: PropTypes.func.isRequired,
+  values: PropTypes.shape({
     name: PropTypes.string.isRequired,
-    setValue: PropTypes.func.isRequired,
   }).isRequired,
 };
-
-export default observer(OrgFormFields);
