@@ -1,21 +1,22 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import { localize } from '../../utils/formatters';
-import { connect } from '../../utils/helpers';
-import { projectDevicesPath } from '../../utils/urls';
-import { Button, IconButton } from '../ui/Buttons';
-import { FlexBetween } from '../ui/Flexboxes';
-import { ContentHeader } from '../ui/Headers';
-import { MoreIcon } from '../ui/Icons';
-import { Link } from '../ui/Links';
-import List from '../ui/List';
-import Menu from '../ui/Menu';
-import { Content } from '../ui/Sections';
-import { ContentTitle, Text } from '../ui/Texts';
-import DeleteProjectDialog from './DeleteProjectDialog';
-import EditProjectForm from './EditProjectForm';
+import PropTypes from "prop-types";
+import React, { useContext } from "react";
+import { localize } from "../../utils";
+import Context from "../Context";
+import { Button, IconButton } from "../ui/Buttons";
+import { FlexBetween } from "../ui/Flexboxes";
+import { ContentHeader } from "../ui/Headers";
+import { MoreIcon } from "../ui/Icons";
+import { Link } from "../ui/Links";
+import List from "../ui/List";
+import Menu from "../ui/Menu";
+import { Content } from "../ui/Sections";
+import { ContentTitle, Text } from "../ui/Texts";
+import DeleteProjectDialog from "./DeleteProjectDialog";
+import EditProjectForm from "./EditProjectForm";
 
-function ProjectContent({ openDialog, openDrawer, project }) {
+export default function ProjectContent({ onDelete, onUpdate, project }) {
+  const { openDialog, openDrawer } = useContext(Context);
+
   const name = project.name.trim();
   return (
     <Content as="section">
@@ -23,8 +24,10 @@ function ProjectContent({ openDialog, openDrawer, project }) {
         <FlexBetween>
           <div>
             <ContentTitle gray={!name}>
-              <Link to={projectDevicesPath(project.id)}>
-                {name || 'Untitled project'}
+              <Link
+                to={`/orgs/${project.orgId}/projects/${project.id}/devices`}
+              >
+                {name || "Untitled project"}
               </Link>
             </ContentTitle>
             <Text gray>{project.description}</Text>
@@ -38,14 +41,18 @@ function ProjectContent({ openDialog, openDrawer, project }) {
           >
             <Button
               onClick={() => {
-                openDrawer(<EditProjectForm project={project} />);
+                openDrawer(
+                  <EditProjectForm project={project} onUpdate={onUpdate} />,
+                );
               }}
             >
               Edit
             </Button>
             <Button
               onClick={() => {
-                openDialog(<DeleteProjectDialog project={project} />);
+                openDialog(
+                  <DeleteProjectDialog project={project} onDelete={onDelete} />,
+                );
               }}
             >
               Delete
@@ -55,8 +62,8 @@ function ProjectContent({ openDialog, openDrawer, project }) {
       </ContentHeader>
       <List
         items={[
-          ['ID', project.id],
-          ['Date Created', localize(project.createdAt)],
+          ["ID", project.id],
+          ["Date Created", localize(project.createdAt)],
         ]}
       />
     </Content>
@@ -64,20 +71,13 @@ function ProjectContent({ openDialog, openDrawer, project }) {
 }
 
 ProjectContent.propTypes = {
-  openDialog: PropTypes.func.isRequired,
-  openDrawer: PropTypes.func.isRequired,
+  onDelete: PropTypes.func.isRequired,
+  onUpdate: PropTypes.func.isRequired,
   project: PropTypes.shape({
     createdAt: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
     id: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired,
+    orgId: PropTypes.string.isRequired,
   }).isRequired,
 };
-
-export default connect(
-  ProjectContent,
-  ({ uiStore }) => ({
-    openDialog: uiStore.openDialog,
-    openDrawer: uiStore.openDrawer,
-  }),
-);
