@@ -1,36 +1,37 @@
-import { observer } from 'mobx-react';
-import PropTypes from 'prop-types';
-import React from 'react';
-import { Input, Label, ReadOnlyInput } from '../ui/Forms';
-import Select from '../ui/Select';
+import PropTypes from "prop-types";
+import React from "react";
+import { Input, Label, ReadOnlyInput } from "../ui/Forms";
+import Select from "../ui/Select";
 
-function ProjectFormFields({ form, org, selectOrgFrom }) {
+export default function ProjectFormFields({ org, setValue, values }) {
   return (
     <React.Fragment>
       <Label htmlFor="orgId">
         Organization
-        {org ? (
-          <ReadOnlyInput id="orgId" value={org.name} />
+        {typeof org === "string" ? (
+          <ReadOnlyInput id="orgId" value={org} />
         ) : (
-          <Select id="orgId" value={form.orgId} onChange={form.setValue}>
-            {selectOrgFrom.map(option => (
-              <option key={option.id} value={option.id}>
-                {option.name}
-              </option>
-            ))}
+          <Select id="orgId" value={values.orgId} onChange={setValue}>
+            {[...org]
+              .sort((a, b) => a.name.localeCompare(b.name))
+              .map(option => (
+                <option key={option.id} value={option.id}>
+                  {option.name}
+                </option>
+              ))}
           </Select>
         )}
       </Label>
       <Label htmlFor="name">
         Name
-        <Input id="name" value={form.name} onChange={form.setValue} />
+        <Input id="name" value={values.name} onChange={setValue} />
       </Label>
       <Label htmlFor="description">
         Description
         <Input
           id="description"
-          value={form.description}
-          onChange={form.setValue}
+          value={values.description}
+          onChange={setValue}
         />
       </Label>
     </React.Fragment>
@@ -38,26 +39,19 @@ function ProjectFormFields({ form, org, selectOrgFrom }) {
 }
 
 ProjectFormFields.propTypes = {
-  form: PropTypes.shape({
+  org: PropTypes.oneOfType([
+    PropTypes.arrayOf(
+      PropTypes.shape({
+        id: PropTypes.string.isRequired,
+        name: PropTypes.string.isRequired,
+      }),
+    ),
+    PropTypes.string,
+  ]).isRequired,
+  setValue: PropTypes.func.isRequired,
+  values: PropTypes.shape({
     description: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired,
     orgId: PropTypes.string.isRequired,
-    setValue: PropTypes.func.isRequired,
   }).isRequired,
-  org: PropTypes.shape({
-    name: PropTypes.string.isRequired,
-  }),
-  selectOrgFrom: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
-    }),
-  ),
 };
-
-ProjectFormFields.defaultProps = {
-  org: null,
-  selectOrgFrom: [],
-};
-
-export default observer(ProjectFormFields);
