@@ -1,11 +1,12 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import styled from 'styled-components';
-import { connect } from '../../utils/helpers';
-import { Button } from './Buttons';
-import { FlexBetween } from './Flexboxes';
-import { ArrowDownIcon } from './Icons';
-import Menu from './Menu';
+import jwtDecode from "jwt-decode";
+import PropTypes from "prop-types";
+import React, { useContext } from "react";
+import styled from "styled-components";
+import Context from "../Context";
+import { Button } from "./Buttons";
+import { FlexBetween } from "./Flexboxes";
+import { ArrowDownIcon } from "./Icons";
+import Menu from "./Menu";
 
 const ControlButton = styled(Button)`
   text-transform: none;
@@ -18,14 +19,16 @@ const ControlIcon = styled(ArrowDownIcon)`
   width: var(--size-16);
 `;
 
-function AppBar({ children, signOut, user }) {
+export default function AppBar({ children }) {
+  const { auth, signOut } = useContext(Context);
+
   return (
     <FlexBetween>
       {children}
       <Menu
         control={
           <ControlButton>
-            {user.email} <ControlIcon aria-hidden />
+            {jwtDecode(auth).email} <ControlIcon aria-hidden />
           </ControlButton>
         }
       >
@@ -40,16 +43,4 @@ AppBar.propTypes = {
     PropTypes.element,
     PropTypes.arrayOf(PropTypes.element),
   ]).isRequired,
-  signOut: PropTypes.func.isRequired,
-  user: PropTypes.shape({
-    email: PropTypes.string.isRequired,
-  }).isRequired,
 };
-
-export default connect(
-  AppBar,
-  ({ userStore }) => ({
-    signOut: userStore.signOut,
-    user: userStore.user,
-  }),
-);
