@@ -1,7 +1,6 @@
-import { observer } from 'mobx-react';
-import React from 'react';
-import styled from 'styled-components';
-import { Link } from './Links';
+import React from "react";
+import styled from "styled-components";
+import { Link } from "./Links";
 
 const Wrapper = styled.footer`
   margin-bottom: var(--size-32);
@@ -25,7 +24,7 @@ const Item = styled.li`
   }
 `;
 
-function Footer() {
+export default function Footer() {
   return (
     <Wrapper>
       <Resources>
@@ -47,5 +46,3 @@ function Footer() {
     </Wrapper>
   );
 }
-
-export default observer(Footer);
