@@ -1,30 +1,26 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import styled from 'styled-components';
-import { summarize } from '../../utils/formatters';
-import { connect } from '../../utils/helpers';
-import Banner from './Banner';
-import { IconButton, RaisedButton } from './Buttons';
-import { FlexEnd, FlexStart } from './Flexboxes';
-import { Form } from './Forms';
-import { PageHeader } from './Headers';
-import { CloseIcon } from './Icons';
-import { Drawer } from './Modals';
-import { PageTitle } from './Texts';
-import { Width320 } from './Widths';
+import PropTypes from "prop-types";
+import React, { useContext, useState } from "react";
+import styled from "styled-components";
+import Context from "../Context";
+import { IconButton, RaisedButton } from "./Buttons";
+import { FlexEnd, FlexStart } from "./Flexboxes";
+import FocusTrap from "./FocusTrap";
+import { Form } from "./Forms";
+import { PageHeader } from "./Headers";
+import { CloseIcon } from "./Icons";
+import { Drawer } from "./Modals";
+import { PageTitle } from "./Texts";
+import { Width320 } from "./Widths";
 
 const Title = styled(PageTitle)`
   margin-left: var(--size-16);
 `;
 
-function FormDrawer({
-  buttonText,
-  children,
-  closeDrawer,
-  onSubmit,
-  openBanner,
-  title,
-}) {
+export default function FormDrawer({ buttonText, children, onSubmit, title }) {
+  const { closeDrawer, errorBoundary } = useContext(Context);
+
+  const [isSubmitting, setSubmitting] = useState(false);
+
   return (
     <Drawer onRequestClose={closeDrawer} contentLabel={title}>
       <PageHeader>
@@ -42,11 +38,13 @@ function FormDrawer({
         <Form
           onSubmit={async event => {
             event.preventDefault();
-            try {
+            setSubmitting(true);
+            const error = await errorBoundary(async () => {
               await onSubmit();
               closeDrawer();
-            } catch (error) {
-              openBanner(<Banner>{summarize(error)}</Banner>);
+            });
+            if (error && error.response.status !== 408) {
+              setSubmitting(false);
             }
           }}
         >
@@ -56,6 +54,7 @@ function FormDrawer({
           </FlexEnd>
         </Form>
       </Width320>
+      {isSubmitting && <FocusTrap label={`Processing ${title} request`} />}
     </Drawer>
   );
 }
@@ -66,16 +65,6 @@ FormDrawer.propTypes = {
     PropTypes.element,
     PropTypes.arrayOf(PropTypes.element),
   ]).isRequired,
-  closeDrawer: PropTypes.func.isRequired,
   onSubmit: PropTypes.func.isRequired,
-  openBanner: PropTypes.func.isRequired,
   title: PropTypes.string.isRequired,
 };
-
-export default connect(
-  FormDrawer,
-  ({ uiStore }) => ({
-    closeDrawer: uiStore.closeDrawer,
-    openBanner: uiStore.openBanner,
-  }),
-);
