@@ -1,7 +1,7 @@
-import styled from 'styled-components';
+import styled from "styled-components";
 
 export const Form = styled.form.attrs({
-  autoComplete: 'off',
+  autoComplete: "off",
   noValidate: true,
 })`
   background-color: var(--color-white);
@@ -17,8 +17,8 @@ export const Label = styled.label`
 `;
 
 export const Input = styled.input.attrs({
-  autoComplete: 'off',
-  type: 'text',
+  autoComplete: "off",
+  type: "text",
 })`
   border: 1px solid var(--color-medium-gray);
   border-radius: var(--corner-radius);
