@@ -1,5 +1,5 @@
-import styled from 'styled-components';
-import LogoSvg from './images/logo.svg';
+import styled from "styled-components";
+import LogoSvg from "../../icons/logo.svg";
 
 export default styled(LogoSvg)`
   display: block;
