@@ -1,33 +1,21 @@
-import { Provider } from 'mobx-react';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import ReactModal from 'react-modal';
-import App from './components/App';
-import Api from './models/Api';
-import Stores from './models/Stores';
+import React from "react";
+import ReactDOM from "react-dom";
+import ReactModal from "react-modal";
+import { setBaseUrl } from "./api";
+import App from "./components/App";
 
-const stores = Stores.create(
-  {},
-  {
-    api: new Api(),
-  },
-);
+setBaseUrl("/api");
 
 function renderToDom() {
-  const appElement = document.getElementById('app');
+  const appElement = document.getElementById("app");
   ReactModal.setAppElement(appElement);
-  ReactDOM.render(
-    <Provider stores={stores}>
-      <App />
-    </Provider>,
-    appElement,
-  );
+  ReactDOM.render(<App />, appElement);
 }
 
 renderToDom();
 
 if (module.hot) {
-  module.hot.accept('./components/App', () => {
+  module.hot.accept("./components/App", () => {
     renderToDom();
   });
 }
